+++ betterrandom/src/main/java/io/github/pr0methean/betterrandom/prng/Pcg64Random.java
@@ -18,8 +18,7 @@ import java.util.concurrent.atomic.AtomicLong;
  * href="http://www.pcg-random.org/">http://www.pcg-random.org/</a>. Period is 2<sup>62</sup> bits.
  * This PRNG is seekable.
  * </p><p>
- * Concurrency is lockless, but contention is still possible due to a retry loop of {@link
- * AtomicLong#compareAndSet(long, long)}. Thus, sharing a single instance across threads isn't
+ * Sharing a single instance across threads that are frequently using it concurrently isn't
  * recommended unless memory is too constrained to use with a {@link ThreadLocalRandomWrapper}.
  * </p>
  * @author M.E. O'Neill (algorithm and C++ implementation)
@@ -35,7 +34,7 @@ public class Pcg64Random extends BaseRandom implements SeekableRandom {
   private static final int ROTATION2 = (Long.SIZE - Integer.SIZE - WANTED_OP_BITS);
   private static final int MASK = (1 << WANTED_OP_BITS) - 1;
 
-  private AtomicLong internal;
+  private final AtomicLong internal;
 
   public Pcg64Random() {
     this(DefaultSeedGenerator.DEFAULT_SEED_GENERATOR);
@@ -58,6 +57,24 @@ public class Pcg64Random extends BaseRandom implements SeekableRandom {
     internal = new AtomicLong(seed);
   }
 
+  @Override protected long nextLongNoEntropyDebit() {
+    lock.lock();
+    try {
+      return ((long) (next(32)) << 32) + next(32);
+    } finally {
+      lock.unlock();
+    }
+  }
+
+  @Override public double nextDoubleNoEntropyDebit() {
+    lock.lock();
+    try {
+      return super.nextDoubleNoEntropyDebit();
+    } finally {
+      lock.unlock();
+    }
+  }
+
   @Override public byte[] getSeed() {
     return BinaryUtils.convertLongToBytes(internal.get()).clone();
   }
@@ -65,12 +82,20 @@ public class Pcg64Random extends BaseRandom implements SeekableRandom {
   @SuppressWarnings("NonSynchronizedMethodOverridesSynchronizedMethod") @Override
   public void setSeed(long seed) {
     if (internal != null) {
-      internal.set(seed);
+      lock.lock();
+      try {
+        internal.set(seed);
+      } finally {
+        lock.unlock();
+      }
     }
     creditEntropyForNewSeed(Long.BYTES);
   }
 
   @Override public void advance(long delta) {
+    if (delta == 0) {
+      return;
+    }
     // The method used here is based on Brown, "Random Number Generation
     // with Arbitrary Stride,", Transactions of the American Nuclear
     // Society (Nov. 1994).  The algorithm is very similar to fast
@@ -90,7 +115,12 @@ public class Pcg64Random extends BaseRandom implements SeekableRandom {
     }
     final long finalAccMult = accMult;
     final long finalAccPlus = accPlus;
-    internal.updateAndGet(old -> (finalAccMult * old) + finalAccPlus);
+    lock.lock();
+    try {
+      internal.updateAndGet(old -> (finalAccMult * old) + finalAccPlus);
+    } finally {
+      lock.unlock();
+    }
   }
 
   @Override public void setSeedInternal(byte[] seed) {
@@ -99,17 +129,21 @@ public class Pcg64Random extends BaseRandom implements SeekableRandom {
       throw new IllegalArgumentException("Pcg64Random requires an 8-byte seed");
     }
     if (internal != null) {
-      internal.set(BinaryUtils.convertBytesToLong(seed));
+      lock.lock();
+      try {
+        internal.set(BinaryUtils.convertBytesToLong(seed));
+      } finally {
+        lock.unlock();
+      }
     }
   }
 
   @Override protected int next(int bits) {
-    internal.updateAndGet(old -> (MULTIPLIER * old) + INCREMENT);
     long oldInternal;
     long newInternal;
     do {
       oldInternal = internal.get();
-      newInternal = oldInternal;
+      newInternal = oldInternal * MULTIPLIER + INCREMENT;
       newInternal ^= oldInternal >>> ROTATION1;
     } while (!internal.compareAndSet(oldInternal, newInternal));
     int rot = (int) (oldInternal >>> (Long.SIZE - WANTED_OP_BITS)) & MASK;
@@ -124,6 +158,15 @@ public class Pcg64Random extends BaseRandom implements SeekableRandom {
 
   @Override protected ToStringHelper addSubclassFields(ToStringHelper original) {
     return original.add("internal", internal.get());
+  }
+
+  @Override public double nextGaussian() {
+    lock.lock();
+    try {
+      return super.nextGaussian();
+    } finally {
+      lock.unlock();
+    }
   }
 
   @Override public int getNewSeedLength() {