+++ betterrandom/src/main/java/io/github/pr0methean/betterrandom/prng/adapter/BaseSplittableRandomAdapter.java
@@ -29,10 +29,6 @@ public abstract class BaseSplittableRandomAdapter extends BaseRandom {
     super(seed);
   }
 
-  @Override protected void setSeedInternal(final byte[] seed) {
-    super.setSeedInternal(seed);
-  }
-
   /**
    * Returns the {@link SplittableRandom} that is to be used to generate random numbers for the
    * current thread. ({@link SplittableRandom} isn't thread-safe.) Called by all the {@code next*}
@@ -44,7 +40,7 @@ public abstract class BaseSplittableRandomAdapter extends BaseRandom {
   /** Delegates to {@link SplittableRandom#nextDouble(double) SplittableRandom.nextDouble(bound)}. */
   @Override public double nextDouble(final double bound) {
     final double out = getSplittableRandom().nextDouble(bound);
-    recordEntropySpent(ENTROPY_OF_DOUBLE);
+    debitEntropy(ENTROPY_OF_DOUBLE);
     return out;
   }
 
@@ -54,13 +50,13 @@ public abstract class BaseSplittableRandomAdapter extends BaseRandom {
    */
   @Override public double nextDouble(final double origin, final double bound) {
     final double out = getSplittableRandom().nextDouble(origin, bound);
-    recordEntropySpent(ENTROPY_OF_DOUBLE);
+    debitEntropy(ENTROPY_OF_DOUBLE);
     return out;
   }
 
   /** Delegates to {@link SplittableRandom#nextInt()} or {@link SplittableRandom#nextInt(int)}. */
   @Override protected int next(final int bits) {
-    recordEntropySpent(bits);
+    debitEntropy(bits);
     return (bits >= 32) ? getSplittableRandom().nextInt()
         : getSplittableRandom().nextInt(1 << (bits - 1));
   }
@@ -72,20 +68,20 @@ public abstract class BaseSplittableRandomAdapter extends BaseRandom {
     for (int i = 0; i < bytes.length; i++) {
       bytes[i] = (byte) (local.nextInt(256));
     }
-    recordEntropySpent(bytes.length * (long) (Byte.SIZE));
+    debitEntropy(bytes.length * (long) (Byte.SIZE));
   }
 
   /** Delegates to {@link SplittableRandom#nextInt()}. */
   @Override public int nextInt() {
     final int out = getSplittableRandom().nextInt();
-    recordEntropySpent(Integer.SIZE);
+    debitEntropy(Integer.SIZE);
     return out;
   }
 
   /** Delegates to {@link SplittableRandom#nextInt(int) SplittableRandom.nextInt(bound)}. */
   @Override public int nextInt(final int bound) {
     final int out = getSplittableRandom().nextInt(bound);
-    recordEntropySpent(entropyOfInt(0, bound));
+    debitEntropy(entropyOfInt(0, bound));
     return out;
   }
 
@@ -95,7 +91,7 @@ public abstract class BaseSplittableRandomAdapter extends BaseRandom {
    */
   @Override public int nextInt(final int origin, final int bound) {
     final int out = getSplittableRandom().nextInt(origin, bound);
-    recordEntropySpent(entropyOfInt(origin, bound));
+    debitEntropy(entropyOfInt(origin, bound));
     return out;
   }
 
@@ -103,7 +99,7 @@ public abstract class BaseSplittableRandomAdapter extends BaseRandom {
   @Override protected boolean withProbabilityInternal(final double probability) {
     final boolean result = getSplittableRandom().nextDouble() < probability;
     // We're only outputting one bit
-    recordEntropySpent(1);
+    debitEntropy(1);
     return result;
   }
 
@@ -119,7 +115,7 @@ public abstract class BaseSplittableRandomAdapter extends BaseRandom {
   /** Delegates to {@link SplittableRandom#nextLong(long) SplittableRandom.nextLong(bound)}. */
   @Override public long nextLong(final long bound) {
     final long out = getSplittableRandom().nextLong(bound);
-    recordEntropySpent(entropyOfLong(0, bound));
+    debitEntropy(entropyOfLong(0, bound));
     return out;
   }
 
@@ -129,15 +125,13 @@ public abstract class BaseSplittableRandomAdapter extends BaseRandom {
    */
   @Override public long nextLong(final long origin, final long bound) {
     final long out = getSplittableRandom().nextLong(origin, bound);
-    recordEntropySpent(entropyOfLong(origin, bound));
+    debitEntropy(entropyOfLong(origin, bound));
     return out;
   }
 
   /** Delegates to {@link SplittableRandom#nextDouble()}. */
-  @Override public double nextDouble() {
-    final double out = getSplittableRandom().nextDouble();
-    recordEntropySpent(ENTROPY_OF_DOUBLE);
-    return out;
+  @Override protected double nextDoubleNoEntropyDebit() {
+    return getSplittableRandom().nextDouble();
   }
 
   /**
@@ -146,15 +140,23 @@ public abstract class BaseSplittableRandomAdapter extends BaseRandom {
   @Override public double nextGaussian() {
     // Upper bound. 2 Gaussians are generated from 2 nextDouble calls, which once made are either
     // used or rerolled.
-    recordEntropySpent(ENTROPY_OF_DOUBLE);
+    debitEntropy(ENTROPY_OF_DOUBLE);
 
     return internalNextGaussian(() -> getSplittableRandom().nextDouble());
+  }
+
+  @Override protected void lockForNextGaussian() {
+    // No-op.
+  }
+
+  @Override protected void unlockForNextGaussian() {
+    // No-op.
   }
 
   /** Delegates to {@link SplittableRandom#nextBoolean()}. */
   @Override public boolean nextBoolean() {
     final boolean out = getSplittableRandom().nextBoolean();
-    recordEntropySpent(1);
+    debitEntropy(1);
     return out;
   }
 
@@ -162,7 +164,7 @@ public abstract class BaseSplittableRandomAdapter extends BaseRandom {
   @Override public float nextFloat() {
     final float out =
         getSplittableRandom().nextInt(1 << ENTROPY_OF_FLOAT) / ((float) (1 << ENTROPY_OF_FLOAT));
-    recordEntropySpent(ENTROPY_OF_FLOAT);
+    debitEntropy(ENTROPY_OF_FLOAT);
     return out;
   }