+++ betterrandom/src/main/java/io/github/pr0methean/betterrandom/seed/RandomSeederThread.java
@@ -4,9 +4,6 @@ import io.github.pr0methean.betterrandom.ByteArrayReseedableRandom;
 import io.github.pr0methean.betterrandom.EntropyCountingRandom;
 import io.github.pr0methean.betterrandom.util.LogPreFormatter;
 import io.github.pr0methean.betterrandom.util.LooperThread;
-import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -20,6 +17,7 @@ import java.util.WeakHashMap;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.locks.Condition;
 import java.util.logging.Level;
 
@@ -35,25 +33,16 @@ public final class RandomSeederThread extends LooperThread {
   private static final LogPreFormatter LOG = new LogPreFormatter(RandomSeederThread.class);
   @SuppressWarnings("StaticCollection") private static final Map<SeedGenerator, RandomSeederThread>
       INSTANCES = Collections.synchronizedMap(new WeakHashMap<>(1));
-  private static final long serialVersionUID = 5229976461051217528L;
   private static final long POLL_INTERVAL = 60;
+  private static final AtomicBoolean loggingEnabled = new AtomicBoolean(true);
   private final SeedGenerator seedGenerator;
   private final byte[] longSeedArray = new byte[8];
-  // WeakHashMap-based Set can't be serialized, so read & write this copy instead
-  private final Set<Random> prngsSerial = new HashSet<>();
   private transient Set<Random> prngs;
   private transient ByteBuffer longSeedBuffer;
   private transient Condition waitWhileEmpty;
   private transient Condition waitForEntropyDrain;
   private transient Set<Random> prngsThisIteration;
   private transient WeakHashMap<ByteArrayReseedableRandom, byte[]> seedArrays;
-
-  public RandomSeederThread(final ThreadGroup group, final Runnable target, final String name,
-      final long stackSize, final SeedGenerator seedGenerator) {
-    super(group, target, name, stackSize);
-    this.seedGenerator = seedGenerator;
-    initTransientFields();
-  }
 
   /**
    * Private constructor because only one instance per seed source.
@@ -64,19 +53,26 @@ public final class RandomSeederThread extends LooperThread {
   }
 
   /**
+   * Controls whether or not to log new instance creation and errors.
+   * @param enabled whether to enable logging
+   */
+  public static void setLoggingEnabled(boolean enabled) {
+    loggingEnabled.set(enabled);
+  }
+
+  /**
    * Obtain the instance for the given {@link SeedGenerator}, creating and starting it if it doesn't
    * exist.
    * @param seedGenerator the {@link SeedGenerator} to use to seed PRNGs registered with this
    *     RandomSeederThread.
    * @return a RandomSeederThread that is running and is backed by {@code seedGenerator}.
-   * @deprecated Relying on a specific instance causes issues if the {@link SeedGenerator} can ever
-   * throw a {@link SeedException}.
    */
-  @Deprecated
-  public static RandomSeederThread getInstance(final SeedGenerator seedGenerator) {
+  private static RandomSeederThread getInstance(final SeedGenerator seedGenerator) {
     synchronized (INSTANCES) {
       return INSTANCES.computeIfAbsent(seedGenerator, seedGen -> {
-        LOG.info("Creating a RandomSeederThread for %s", seedGen);
+        if (loggingEnabled.get()) {
+          LOG.info("Creating a RandomSeederThread for %s", seedGen);
+        }
         final RandomSeederThread thread = new RandomSeederThread(seedGen);
         thread.setName("RandomSeederThread for " + seedGen);
         thread.setDaemon(true);
@@ -84,6 +80,18 @@ public final class RandomSeederThread extends LooperThread {
         thread.start();
         return thread;
       });
+    }
+  }
+
+  /**
+   * Returns whether a RandomSeederThread using the given {@link SeedGenerator} is running or not.
+   * @param seedGenerator a {@link SeedGenerator} to find an instance for.
+   * @return true if a RandomSeederThread using the given {@link SeedGenerator} is running; false
+   *     otherwise.
+   */
+  public static boolean hasInstance(final SeedGenerator seedGenerator) {
+    synchronized (INSTANCES) {
+      return INSTANCES.containsKey(seedGenerator);
     }
   }
 
@@ -109,44 +117,6 @@ public final class RandomSeederThread extends LooperThread {
     } while (!toStop.isEmpty());
   }
 
-  private void initTransientFields() {
-    prngs = Collections.synchronizedSet(Collections.newSetFromMap(new WeakHashMap<>(1)));
-    longSeedBuffer = ByteBuffer.wrap(longSeedArray);
-    waitWhileEmpty = lock.newCondition();
-    waitForEntropyDrain = lock.newCondition();
-    prngsThisIteration = new HashSet<>(1);
-    seedArrays = new WeakHashMap<>(1);
-  }
-
-  @Override protected Object readResolve() {
-    return getInstance(seedGenerator);
-  }
-
-  private void readObject(final ObjectInputStream in) throws IOException, ClassNotFoundException {
-    in.defaultReadObject();
-    initTransientFields();
-    if (!prngsSerial.isEmpty()) {
-      synchronized (prngs) {
-        prngs.addAll(prngsSerial);
-      }
-      lock.lock();
-      try {
-        waitWhileEmpty.signalAll();
-      } finally {
-        lock.unlock();
-      }
-      prngsSerial.clear();
-    }
-  }
-
-  private void writeObject(final ObjectOutputStream out) throws IOException {
-    synchronized (prngs) {
-      prngsSerial.addAll(prngs);
-    }
-    out.defaultWriteObject();
-    prngsSerial.clear();
-  }
-
   /**
    * Asynchronously triggers reseeding of the given {@link EntropyCountingRandom} if it is
    * associated with a live RandomSeederThread corresponding to the given {@link SeedGenerator}.
@@ -160,15 +130,62 @@ public final class RandomSeederThread extends LooperThread {
     }
   }
 
+  public static boolean isEmpty(SeedGenerator seedGenerator) {
+    synchronized (INSTANCES) {
+      return (!hasInstance(seedGenerator)) || getInstance(seedGenerator).isEmpty();
+    }
+  }
+
+  /**
+   * Add one or more {@link Random} instances to the thread for the given {@link SeedGenerator}.
+   * @param seedGenerator The {@link SeedGenerator} that will reseed the {@code randoms}
+   * @param randoms One or more {@link Random} instances to be reseeded
+   */
+  public static void add(SeedGenerator seedGenerator, final Random... randoms) {
+    synchronized (INSTANCES) {
+      getInstance(seedGenerator).add(randoms);
+    }
+  }
+
+  /**
+   * Remove one or more {@link Random} instances from the thread for the given {@link SeedGenerator}
+   * if such a thread exists and contains them.
+   * @param seedGenerator The {@link SeedGenerator} that will reseed the {@code randoms}
+   * @param randoms One or more {@link Random} instances to be reseeded
+   */
+  public static void remove(SeedGenerator seedGenerator, final Random... randoms) {
+    synchronized (INSTANCES) {
+      RandomSeederThread thread = INSTANCES.get(seedGenerator);
+      if (thread != null) {
+        thread.remove(randoms);
+      }
+    }
+  }
+
+  public static void stopIfEmpty(SeedGenerator seedGenerator) {
+    synchronized (INSTANCES) {
+      if (hasInstance(seedGenerator)) {
+        getInstance(seedGenerator).stopIfEmpty();
+      }
+    }
+  }
+
+  private void initTransientFields() {
+    prngs = Collections.synchronizedSet(Collections.newSetFromMap(new WeakHashMap<>(1)));
+    longSeedBuffer = ByteBuffer.wrap(longSeedArray);
+    waitWhileEmpty = lock.newCondition();
+    waitForEntropyDrain = lock.newCondition();
+    prngsThisIteration = new HashSet<>(1);
+    seedArrays = new WeakHashMap<>(1);
+  }
+
   /**
    * Asynchronously triggers reseeding of the given {@link EntropyCountingRandom} if it is
    * associated with a live RandomSeederThread.
    * @param random a {@link Random} object.
    * @return Whether or not the reseed was successfully scheduled.
-   * @deprecated Causes weird race conditions
    */
-  @Deprecated
-  @SuppressWarnings("UnusedReturnValue") public boolean asyncReseed(final Random random) {
+  private boolean asyncReseed(final Random random) {
     if (!(random instanceof EntropyCountingRandom)) {
       // Reseed of non-entropy-counting Random happens every iteration anyway
       return prngs.contains(random);
@@ -193,7 +210,7 @@ public final class RandomSeederThread extends LooperThread {
   }
 
   @SuppressWarnings({"InfiniteLoopStatement", "ObjectAllocationInLoop", "AwaitNotInLoop"}) @Override
-  public boolean iterate() throws InterruptedException {
+  protected boolean iterate() throws InterruptedException {
     while (true) {
       synchronized (prngs) {
         prngsThisIteration.addAll(prngs);
@@ -230,8 +247,10 @@ public final class RandomSeederThread extends LooperThread {
       } catch (final Throwable t) {
         // Must unlock before interrupt; otherwise we somehow get a deadlock
         lock.unlock();
-        LOG.error("%s", t);
-        LOG.logStackTrace(Level.SEVERE, t.getStackTrace());
+        if (loggingEnabled.get()) {
+          LOG.error("%s", t);
+          LOG.logStackTrace(Level.SEVERE, t.getStackTrace());
+        }
         interrupt();
         // Must lock again before returning, so we can notify conditions
         lock.lock();
@@ -256,20 +275,9 @@ public final class RandomSeederThread extends LooperThread {
    * Returns true if no {@link Random} instances are registered with this RandomSeederThread.
    * @return true if no {@link Random} instances are registered with this RandomSeederThread.
    */
-  public boolean isEmpty() {
+  private boolean isEmpty() {
     synchronized (prngs) {
       return prngs.isEmpty();
-    }
-  }
-
-  /**
-   * Add one or more {@link Random} instances to the thread for the given {@link SeedGenerator}.
-   * @param seedGenerator The {@link SeedGenerator} that will reseed the {@code randoms}
-   * @param randoms One or more {@link Random} instances to be reseeded
-   */
-  public static void add(SeedGenerator seedGenerator, final Random... randoms) {
-    synchronized (INSTANCES) {
-      getInstance(seedGenerator).add(randoms);
     }
   }
 
@@ -279,10 +287,8 @@ public final class RandomSeederThread extends LooperThread {
    * ByteArrayReseedableRandom#setSeed(byte[])}, as one of those methods may be called immediately
    * and this would cause a circular deadlock.
    * @param randoms One or more {@link Random} instances to be reseeded.
-   * @deprecated Causes weird race conditions.
    */
-  @Deprecated
-  public void add(final Random... randoms) {
+  private void add(final Random... randoms) {
     lock.lock();
     if (getState() == State.TERMINATED || isInterrupted()) {
       throw new IllegalStateException("Already shut down");
@@ -297,51 +303,28 @@ public final class RandomSeederThread extends LooperThread {
   }
 
   /**
-   * Remove one or more {@link Random} instances from the thread for the given {@link SeedGenerator}
-   * if such a thread exists and contains them.
-   * @param seedGenerator The {@link SeedGenerator} that will reseed the {@code randoms}
-   * @param randoms One or more {@link Random} instances to be reseeded
-   */
-  public static void remove(SeedGenerator seedGenerator, final Random... randoms) {
-    synchronized (INSTANCES) {
-      RandomSeederThread thread = INSTANCES.get(seedGenerator);
-      if (thread != null) {
-        thread.remove(randoms);
-      }
-    }
-  }
-
-  /**
    * Remove one or more {@link Random} instances. If this is called while {@link #getState()} ==
    * {@link State#RUNNABLE}, they may still be reseeded once more.
    * @param randoms the {@link Random} instances to remove.
-   * @deprecated Causes weird race conditions.
    */
-  @Deprecated
-  public void remove(final Random... randoms) {
+  private void remove(final Random... randoms) {
     prngs.removeAll(Arrays.asList(randoms));
   }
 
   /**
    * Shut down this thread if no {@link Random} instances are registered with it.
    */
-  public void stopIfEmpty() {
+  private void stopIfEmpty() {
     lock.lock();
     try {
       if (isEmpty()) {
-        LOG.info("Stopping empty RandomSeederThread for %s", seedGenerator);
+        if (loggingEnabled.get()) {
+          LOG.info("Stopping empty RandomSeederThread for %s", seedGenerator);
+        }
         interrupt();
       }
     } finally {
       lock.unlock();
     }
-  }
-
-  /**
-   * Use this to transition old methods away from the deprecated instance methods.
-   * @return the {@link SeedGenerator} that this thread uses
-   */
-  public SeedGenerator getSeedGenerator() {
-    return seedGenerator;
   }
 }