+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/prng/ReseedingThreadLocalRandomWrapperTest.java
@@ -1,13 +1,16 @@
 package io.github.pr0methean.betterrandom.prng;
 
+import static io.github.pr0methean.betterrandom.TestUtils.assertGreaterOrEqual;
+import static io.github.pr0methean.betterrandom.seed.DefaultSeedGenerator.DEFAULT_SEED_GENERATOR;
+
+import io.github.pr0methean.betterrandom.prng.RandomTestUtils.EntropyCheckMode;
 import io.github.pr0methean.betterrandom.seed.DefaultSeedGenerator;
-import io.github.pr0methean.betterrandom.seed.FailingSeedGenerator;
 import io.github.pr0methean.betterrandom.seed.SeedException;
-import io.github.pr0methean.betterrandom.seed.SeedGenerator;
 import java.io.Serializable;
-import java.util.Map;
+import java.util.Arrays;
 import java.util.Random;
 import java.util.function.Supplier;
+import org.testng.annotations.Test;
 
 public class ReseedingThreadLocalRandomWrapperTest extends ThreadLocalRandomWrapperTest {
 
@@ -16,12 +19,53 @@ public class ReseedingThreadLocalRandomWrapperTest extends ThreadLocalRandomWrap
         .wrapLegacy(Random::new, DefaultSeedGenerator.DEFAULT_SEED_GENERATOR).nextInt();
   }
 
+  @Override protected EntropyCheckMode getEntropyCheckMode() {
+    return EntropyCheckMode.LOWER_BOUND;
+  }
+
   @Override protected Class<? extends BaseRandom> getClassUnderTest() {
-    return super.getClassUnderTest();
+    return ReseedingThreadLocalRandomWrapper.class;
+  }
+
+  @Override @Test public void testReseeding() {
+    final BaseRandom rng =
+        new ReseedingThreadLocalRandomWrapper(DefaultSeedGenerator.DEFAULT_SEED_GENERATOR,
+            (Serializable & Supplier<BaseRandom>) MersenneTwisterRandom::new);
+    rng.nextLong();
+    try {
+      Thread.sleep(1000);
+    } catch (InterruptedException e) {
+      throw new RuntimeException(e);
+    }
+    final byte[] oldSeed = rng.getSeed();
+    try {
+      byte[] newSeed;
+      do {
+        rng.nextLong();
+        Thread.sleep(10);
+        newSeed = rng.getSeed();
+      } while (Arrays.equals(newSeed, oldSeed));
+      assertGreaterOrEqual(rng.getEntropyBits(), newSeed.length * 8L - 1);
+    } catch (InterruptedException e) {
+      throw new RuntimeException(e);
+    }
+  }
+
+  /** Assertion-free since reseeding may cause divergent output. */
+  @Test(timeOut = 10000) public void testSetSeedLong() {
+    createRng().setSeed(0x0123456789ABCDEFL);
+  }
+
+  /** Test for crashes only, since setSeed is a no-op. */
+  @Override @Test public void testSetSeed() throws SeedException {
+    final BaseRandom prng = createRng();
+    prng.nextLong();
+    prng.setSeed(DEFAULT_SEED_GENERATOR.generateSeed(16));
+    prng.nextLong();
   }
 
   @Override protected BaseRandom createRng() throws SeedException {
-    return new ReseedingThreadLocalRandomWrapper(FailingSeedGenerator.FAILING_SEED_GENERATOR,
+    return new ReseedingThreadLocalRandomWrapper(DefaultSeedGenerator.DEFAULT_SEED_GENERATOR,
         (Serializable & Supplier<BaseRandom>) MersenneTwisterRandom::new);
   }
 }