+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/prng/adapter/SingleThreadSplittableRandomAdapterTest.java
@@ -4,6 +4,7 @@ import static io.github.pr0methean.betterrandom.prng.RandomTestUtils.testEquival
 
 import io.github.pr0methean.betterrandom.prng.BaseRandom;
 import io.github.pr0methean.betterrandom.prng.BaseRandomTest;
+import io.github.pr0methean.betterrandom.prng.RandomTestUtils;
 import io.github.pr0methean.betterrandom.seed.DefaultSeedGenerator;
 import io.github.pr0methean.betterrandom.seed.SeedException;
 import io.github.pr0methean.betterrandom.util.CloneViaSerialization;
@@ -13,6 +14,10 @@ public class SingleThreadSplittableRandomAdapterTest extends BaseRandomTest {
 
   @Override protected Class<? extends BaseRandom> getClassUnderTest() {
     return SingleThreadSplittableRandomAdapter.class;
+  }
+
+  @Override @Test(enabled = false) public void testThreadSafety() {
+    // No-op because this class isn't thread-safe.
   }
 
   /**
@@ -41,6 +46,16 @@ public class SingleThreadSplittableRandomAdapterTest extends BaseRandomTest {
     final BaseSplittableRandomAdapter adapter4 = CloneViaSerialization.clone(adapter2);
     testEquivalence(adapter2, adapter3, 20);
     testEquivalence(adapter2, adapter4, 20);
+  }
+
+  @Override public void testSetSeedLong() throws SeedException {
+    final BaseRandom rng = createRng();
+    final BaseRandom rng2 = createRng();
+    rng.nextLong(); // ensure they won't both be in initial state before reseeding
+    rng.setSeed(0x0123456789ABCDEFL);
+    rng2.setSeed(0x0123456789ABCDEFL);
+    assert RandomTestUtils.testEquivalence(rng, rng2, 20)
+        : "Output mismatch after reseeding with same seed";
   }
 
   @Override @Test public void testNullSeed() {