+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/seed/DefaultSeedGeneratorTest.java
@@ -15,6 +15,8 @@
 // ============================================================================
 package io.github.pr0methean.betterrandom.seed;
 
+import static org.testng.Assert.assertTrue;
+
 import java.security.Permission;
 import org.testng.annotations.Test;
 
@@ -22,10 +24,19 @@ import org.testng.annotations.Test;
  * Unit test for {@link DefaultSeedGenerator}.
  * @author Daniel Dyer
  */
-public class DefaultSeedGeneratorTest {
+public class DefaultSeedGeneratorTest extends AbstractSeedGeneratorTest {
+
+  public DefaultSeedGeneratorTest() {
+    super(DefaultSeedGenerator.DEFAULT_SEED_GENERATOR);
+  }
 
   @Test public void testBasicFunction() throws SeedException {
-    SeedTestUtils.testGenerator(DefaultSeedGenerator.DEFAULT_SEED_GENERATOR);
+    SeedTestUtils.testGenerator(seedGenerator);
+  }
+
+  @Test public void testIsWorthTrying() {
+    // Should always be true
+    assertTrue(seedGenerator.isWorthTrying());
   }
 
   /**
@@ -39,7 +50,7 @@ public class DefaultSeedGeneratorTest {
     try {
       // Don't allow file system or network access.
       System.setSecurityManager(new RestrictedSecurityManager(affectedThread));
-      DefaultSeedGenerator.DEFAULT_SEED_GENERATOR.generateSeed(new byte[4]);
+      seedGenerator.generateSeed(new byte[4]);
       // Should get to here without exceptions.
     } finally {
       // Restore the original security manager so that we don't